backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
requests==2.32.3
pydantic==2.9.2
python-multipart==0.0.12
croniter
Pillow
numpy
